feat: clean up the source code
